pyproject.toml
[project]
name = "stdb"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.10"
dependencies = [
    "loguru>=0.7.3",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "spacetimedb-sdk>=0.7.0",
    "websockets>=15.0.1",
]
